refactor: finish moving to typescript
