+++ src/handlers/flags.ts
@@ -2,13 +2,13 @@ import open from 'open'
 import clipboardy from 'clipboardy'
 import { greenBox } from '../infrastructure/boxen'
 
-const outputText = flags => text => {
+export const outputText = flags => text => {
   const output = flags.clear ? text : greenBox(text)
 
   console.log(output)
 }
 
-const doFlagClipboard = flags => text => {
+export const doFlagClipboard = flags => text => {
   if (flags.clipboard) {
     clipboardy.writeSync(text)
   }
@@ -16,14 +16,14 @@ const doFlagClipboard = flags => text => {
   return text
 }
 
-const doOpen = flags => text => {
+export const doOpen = flags => text => {
   if (flags.open) {
     open(text)
   }
   return text
 }
 
-module.exports = {
+export default {
   outputText,
   doFlagClipboard,
   doOpen,
