+++ src/infrastructure/jimp.ts
@@ -2,11 +2,11 @@ import Jimp from 'jimp'
 
 const extractBitmap = ({ bitmap }) => bitmap
 
-const getBitmap = filePath =>
+export const getBitmap = filePath =>
   Promise.resolve(filePath)
     .then(Jimp.read)
     .then(extractBitmap)
 
-module.exports = {
+export default {
   getBitmap,
 }
