--- src/scanFromFile.js
@@ -1,15 +1,0 @@
-const errorHandlers = require('./handlers/error')
-const { getBitmap } = require('./infrastructure/jimp')
-const readQR = require('./infrastructure/qrcode-reader')
-const { outputText, doFlagClipboard, doOpen } = require('./handlers/flags')
-
-const scanFromFile = (filePath, flags) =>
-  Promise.resolve(filePath)
-    .then(getBitmap)
-    .then(readQR)
-    .then(doFlagClipboard(flags))
-    .then(doOpen(flags))
-    .then(outputText(flags))
-    .catch(errorHandlers.scanFromFile(filePath))
-
-module.exports = scanFromFile
